Trait vapoursynth::component::Component

pub unsafe trait Component {
    fn is_valid(format: Format) -> bool;
}

A trait for possible pixel components.

Safety

Implementing this trait allows retrieving slices of pixel data from the frame for the target type, so the target type must be valid for the given format.

Required methods

fn is_valid(format: Format) -> bool

Returns whether this component is valid for this format.

Loading content...