Struct vapoursynth::core::CoreRef

pub struct CoreRef<'core> { /* fields omitted */ }

A reference to a VapourSynth core.

Methods

impl<'core> CoreRef<'core>

pub fn info(self) -> Info

Returns information about the VapourSynth core.

pub fn get_format(&self, id: FormatID) -> Option<Format<'core>>

Retrieves a registered or preset Format by its id. The id can be of a previously registered format, or one of the PresetFormat.

pub fn register_format(
&self,
color_family: ColorFamily,
sample_type: SampleType,
bits_per_sample: u8,
sub_sampling_w: u8,
sub_sampling_h: u8
) -> Option<Format<'core>>

Registers a custom video format.

Returns None if an invalid format is described.

Registering compat formats is not allowed. Only certain privileged built-in filters are allowed to handle compat formats.

RGB formats are not allowed to be subsampled.

pub fn get_plugin_by_id(
&self,
id: &str
) -> Result<Option<Plugin<'core>>, NulError>

Returns a plugin with the given identifier.

pub fn get_plugin_by_namespace(
&self,
namespace: &str
) -> Result<Option<Plugin<'core>>, NulError>

Returns a plugin with the given namespace.

get_plugin_by_id() should be used instead.

pub fn plugins(&self) -> OwnedMap<'core>

Returns a map containing a list of all loaded plugins.

Keys: meaningless unique strings;

Values: namespace, identifier, and full name, separated by semicolons.

pub fn set_max_cache_size(&self, bytes: i64) -> i64

Sets the maximum size of the framebuffer cache. Returns the new maximum size.

pub fn set_thread_count(&self, threads: i32) -> i32

Sets the number of worker threads for the given core.

If the requested number of threads is zero or lower, the number of hardware threads will be detected and used.

Returns the new thread count.