Macro vapoursynth::export_vapoursynth_plugin

macro_rules! export_vapoursynth_plugin {
    ($metadata:expr, [$($filter:expr),*$(,)*]) => { ... };
}

Exports a VapourSynth plugin from this library.

This macro should be used only once at the top level of the library. The library should have a cdylib crate type.

The first parameter is a Metadata expression containing your plugin's metadata.

Following it is a list of values implementing FilterFunction, those are the filter functions the plugin will export.

Example

export_vapoursynth_plugin! {
    Metadata {
        identifier: "com.example.invert",
        namespace: "invert",
        name: "Invert Example Plugin",
        read_only: true,
    },
    [SampleFilterFunction::new(), OtherFunction::new()]
}