Struct vapoursynth::map::Map

pub struct Map<'elem> { /* fields omitted */ }

A VapourSynth map.

A map contains key-value pairs where the value is zero or more elements of a certain type.

Methods

impl<'elem> Map<'elem>

pub fn is_key_valid(key: &str) -> Result<(), InvalidKeyError>

Checks if the key is valid. Valid keys start with an alphabetic character or an underscore, and contain only alphanumeric characters and underscores.

pub fn clear(&mut self)

Clears the map.

pub fn error(&self) -> Option<Cow<str>>

Returns the error message contained in the map, if any.

pub fn set_error(&mut self, error_message: &str) -> Result<()>

Adds an error message to a map. The map is cleared first.

pub fn key_count(&self) -> usize

Returns the number of keys contained in a map.

pub fn key(&self, index: usize) -> &str

Returns a key from a map.

Panics

Panics if index >= self.key_count().

Important traits for Keys<'map, 'elem>
impl<'map, 'elem> Iterator for Keys<'map, 'elem> type Item = &'map str;

pub fn keys(&self) -> Keys

Returns an iterator over all keys in a map.

pub fn value_count(&self, key: &str) -> Result<usize>

Returns the number of elements associated with a key in a map.

pub fn value_type(&self, key: &str) -> Result<ValueType>

Retrieves a value type from a map.

pub fn delete_key(&mut self, key: &str) -> Result<()>

Deletes the given key.

pub fn touch(&mut self, key: &str, value_type: ValueType) -> Result<()>

Touches the key. That is, if the key exists, nothing happens, otherwise a key is created with no values associated.

pub fn get<'map, T: Value<'map, 'elem>>(&'map self, key: &str) -> Result<T>

Retrieves a property value.

pub fn get_iter<'map, T: Value<'map, 'elem>>(
&'map self,
key: &str
) -> Result<ValueIter<'map, 'elem, T>>

Retrieves an iterator over the map values.

pub fn set<'map, T: Value<'map, 'elem>>(
&'map mut self,
key: &str,
x: &T
) -> Result<()>

Sets a property value.

pub fn append<'map, T: Value<'map, 'elem>>(
&'map mut self,
key: &str,
x: &T
) -> Result<()>

Appends a property value.

pub fn get_int(&self, key: &str) -> Result<i64>

Retrieves an integer from a map.

This function retrieves the first value associated with the key.

pub fn get_int_iter<'map>(
&'map self,
key: &str
) -> Result<ValueIter<'map, 'elem, i64>>

Retrieves integers from a map.

pub fn get_int_array(&self, key: &str) -> Result<&[i64]>

Retrieves an array of integers from a map.

This is faster than iterating over a get_int_iter().

pub fn get_float(&self, key: &str) -> Result<f64>

Retrieves a floating point number from a map.

This function retrieves the first value associated with the key.

pub fn get_float_array(&self, key: &str) -> Result<&[f64]>

Retrieves an array of floating point numbers from a map.

This is faster than iterating over a get_float_iter().

pub fn get_float_iter<'map>(
&'map self,
key: &str
) -> Result<ValueIter<'map, 'elem, f64>>

Retrieves floating point numbers from a map.

pub fn get_data(&self, key: &str) -> Result<&[u8]>

Retrieves data from a map.

This function retrieves the first value associated with the key.

pub fn get_data_iter<'map>(
&'map self,
key: &str
) -> Result<ValueIter<'map, 'elem, &'map [u8]>>

Retrieves data from a map.

pub fn get_node(&self, key: &str) -> Result<Node<'elem>>

Retrieves a node from a map.

This function retrieves the first value associated with the key.

pub fn get_node_iter<'map>(
&'map self,
key: &str
) -> Result<ValueIter<'map, 'elem, Node<'elem>>>

Retrieves nodes from a map.

pub fn get_frame(&self, key: &str) -> Result<FrameRef<'elem>>

Retrieves a frame from a map.

This function retrieves the first value associated with the key.

pub fn get_frame_iter<'map>(
&'map self,
key: &str
) -> Result<ValueIter<'map, 'elem, FrameRef<'elem>>>

Retrieves frames from a map.

pub fn get_function(&self, key: &str) -> Result<Function<'elem>>

Retrieves a function from a map.

This function retrieves the first value associated with the key.

pub fn get_function_iter<'map>(
&'map self,
key: &str
) -> Result<ValueIter<'map, 'elem, Function<'elem>>>

Retrieves functions from a map.

pub fn append_int(&mut self, key: &str, x: i64) -> Result<()>

Appends an integer to a map.

pub fn append_float(&mut self, key: &str, x: f64) -> Result<()>

Appends a floating point number to a map.

pub fn append_data(&mut self, key: &str, x: &[u8]) -> Result<()>

Appends data to a map.

pub fn append_node(&mut self, key: &str, x: &Node<'elem>) -> Result<()>

Appends a node to a map.

pub fn append_frame(&mut self, key: &str, x: &Frame<'elem>) -> Result<()>

Appends a frame to a map.

pub fn append_function(&mut self, key: &str, x: &Function<'elem>) -> Result<()>

Appends a function to a map.

pub fn set_int(&mut self, key: &str, x: i64) -> Result<()>

Sets a property value to an integer.

pub fn set_int_array(&mut self, key: &str, x: &[i64]) -> Result<()>

Sets a property value to an integer array.

This is faster than calling append_int() in a loop.

pub fn set_float(&mut self, key: &str, x: f64) -> Result<()>

Sets a property value to a floating point number.

pub fn set_float_array(&mut self, key: &str, x: &[f64]) -> Result<()>

Sets a property value to a floating point number array.

This is faster than calling append_float() in a loop.

pub fn set_data(&mut self, key: &str, x: &[u8]) -> Result<()>

Sets a property value to data.

pub fn set_node(&mut self, key: &str, x: &Node<'elem>) -> Result<()>

Sets a property value to a node.

pub fn set_frame(&mut self, key: &str, x: &Frame<'elem>) -> Result<()>

Sets a property value to a frame.

pub fn set_function(&mut self, key: &str, x: &Function<'elem>) -> Result<()>

Sets a property value to a function.