Struct vapoursynth::vsscript::Environment

pub struct Environment { /* fields omitted */ }

A wrapper for the VSScript environment.

Methods

impl Environment

pub fn new() -> Result<Self, Error>

Creates an empty script environment.

Useful if it is necessary to set some variable in the script environment before evaluating any scripts.

pub fn from_script(script: &str) -> Result<Self, Error>

Creates a script environment and evaluates a script contained in a string.

pub fn from_file<P: AsRef<Path>>(
path: P,
flags: EvalFlags
) -> Result<Self, Error>

Creates a script environment and evaluates a script contained in a file.

pub fn eval_script(&mut self, script: &str) -> Result<(), Error>

Evaluates a script contained in a string.

pub fn eval_file<P: AsRef<Path>>(
&mut self,
path: P,
flags: EvalFlags
) -> Result<(), Error>

Evaluates a script contained in a file.

pub fn clear(&self)

Clears the script environment.

pub fn get_output(&self, index: i32) -> Result<(Node, Option<Node>), Error>

Retrieves a node from the script environment. A node in the script must have been marked for output with the requested index. The second node, if any, contains the alpha clip.

pub fn clear_output(&self, index: i32) -> Result<(), Error>

Cancels a node set for output. The node will no longer be available to get_output().

pub fn get_core(&self) -> Result<CoreRef, Error>

Retrieves the VapourSynth core that was created in the script environment. If a VapourSynth core has not been created yet, it will be created now, with the default options.

pub fn get_variable(&self, name: &str, map: &mut Map) -> Result<(), Error>

Retrieves a variable from the script environment.

pub fn set_variables(&self, variables: &Map) -> Result<(), Error>

Sets variables in the script environment.

pub fn clear_variable(&self, name: &str) -> Result<(), Error>

Deletes a variable from the script environment.