Struct vapoursynth::frame::FrameRefMut

pub struct FrameRefMut<'core> { /* fields omitted */ }

A reference to a mutable frame.

Implementations

impl<'core> FrameRefMut<'core>

pub fn copy_of(core: CoreRef<'_>, frame: &Frame<'core>) -> Self

Creates a copy of the given frame.

The plane data is copy-on-write, so this isn’t very expensive by itself.

Judging by the underlying implementation, it seems that any valid core can be used.

pub unsafe fn new_uninitialized(
core: CoreRef<'core>,
prop_src: Option<&Frame<'core>>,
format: Format<'core>,
resolution: Resolution
) -> Self

Creates a new frame with uninitialized plane data.

Optionally copies the frame properties from the provided prop_src frame.

Safety

The returned frame contains uninitialized plane data. This should be handled carefully. See the docs for std::mem::uninitialized() for more information.

Panics

Panics if the given resolution has components that don’t fit into an i32.