Module vapoursynth::map

VapourSynth maps.

Structs

Keys

An iterator over the keys of a map.

Map

A VapourSynth map.

MapRef

A reference to a VapourSynth map.

MapRefMut

A reference to a mutable VapourSynth map.

OwnedMap

An owned VapourSynth map.

ValueIter

An iterator over the values associated with a certain key of a map.

Enums

Error

The error type for Map operations.

InvalidKeyError

An error indicating the map key is invalid.

ValueType

An enumeration of all possible value types.

Traits

Value

A trait for values which can be stored in a map.

Type Definitions

Result

A specialized Result type for Map operations.