Trait vapoursynth::map::Value

pub trait Value<'map, 'elem: 'map>: Sized {
    fn get_from_map(map: &'map Map<'elem>, key: &str) -> Result<Self>;

    fn get_iter_from_map(
        map: &'map Map<'elem>, 
        key: &str
    ) -> Result<ValueIter<'map, 'elem, Self>>;

    fn store_in_map(
        map: &'map mut Map<'elem>, 
        key: &str, 
        x: &Self
    ) -> Result<()>;

    fn append_to_map(
        map: &'map mut Map<'elem>, 
        key: &str, 
        x: &Self
    ) -> Result<()>;
}

A trait for values which can be stored in a map.

Required methods

fn get_from_map(map: &'map Map<'elem>, key: &str) -> Result<Self>

Retrieves the value from the map.

fn get_iter_from_map(
map: &'map Map<'elem>,
key: &str
) -> Result<ValueIter<'map, 'elem, Self>>

Retrieves an iterator over the values from the map.

fn store_in_map(map: &'map mut Map<'elem>, key: &str, x: &Self) -> Result<()>

Sets the property value in the map.

fn append_to_map(map: &'map mut Map<'elem>, key: &str, x: &Self) -> Result<()>

Appends the value to the map.

Loading content...